requirements.txt
pyautogui==0.9.54
Pillow==10.4.0
keyboard==0.13.5
cx_Freeze==7.2.0
